requirements.txt
docopt==0.6.2
panflute==1.10.6
PyYAML==3.12
